# tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "CarScene": {
        "additionalProperties": false,
        "properties": {
          "bodies": {
            "items": {
              "$ref": "#/components/schemas/SceneBody"
            },
            "title": "Bodies",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "bodies"
        ],
        "title": "CarScene",
        "type": "object"
      },
...
      "ControllerInfo": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "template": {
            "title": "Template",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "path",
          "template"
        ],
        "title": "ControllerInfo",
        "type": "object"
      },
      "CorridorParams": {
        "additionalProperties": false,
        "properties": {
          "door_rate_per_10m": {
            "default": 0.8,
            "minimum": 0.0,
            "title": "Door Rate Per 10M",
            "type": "number"
          },
          "friction_max": {
            "default": 0.9,
            "exclusiveMinimum": 0.0,
            "title": "Friction Max",
            "type": "number"
          },
          "friction_min": {
            "default": 0.6,
            "exclusiveMinimum": 0.0,
            "title": "Friction Min",
            "type": "number"
          },
          "glass_prob": {
            "default": 0.15,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Glass Prob",
            "type": "number"
          },
          "jog_rate_per_10m": {
            "default": 0.25,
            "minimum": 0.0,
            "title": "Jog Rate Per 10M",
            "type": "number"
          },
          "length_m": {
            "default": 60.0,
            "maximum": 120.0,
            "minimum": 20.0,
            "title": "Length M",
            "type": "number"
          },
          "loop": {
            "default": true,
            "title": "Loop",
            "type": "boolean"
          },
          "niche_rate_per_10m": {
            "default": 1.0,
            "minimum": 0.0,
            "title": "Niche Rate Per 10M",
            "type": "number"
          },
          "object_rate_per_10m": {
            "default": 0.6,
            "minimum": 0.0,
            "title": "Object Rate Per 10M",
            "type": "number"
          },
          "pillar_rate_per_10m": {
            "default": 0.4,
            "minimum": 0.0,
            "title": "Pillar Rate Per 10M",
            "type": "number"
          },
          "seed": {
            "default": 0,
            "title": "Seed",
            "type": "integer"
          },
          "wall_height_m": {
            "default": 2.5,
            "exclusiveMinimum": 0.0,
            "title": "Wall Height M",
            "type": "number"
          },
          "width_max_m": {
            "default": 2.4,
            "maximum": 4.0,
            "title": "Width Max M",
            "type": "number"
          },
          "width_min_m": {
            "default": 1.4,
            "minimum": 0.8,
            "title": "Width Min M",
            "type": "number"
          }
        },
        "title": "CorridorParams",
        "type": "object"
      },
      "CorridorResponse": {
        "additionalProperties": false,
        "properties": {
          "centreline": {
            "items": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "number"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array"
            },
            "title": "Centreline",
            "type": "array"
          },
          "primitives": {
            "items": {
              "$ref": "#/components/schemas/Primitive"
            },
            "title": "Primitives",
            "type": "array"
          },
          "track": {
            "additionalProperties": true,
            "title": "Track",
            "type": "object"
          }
        },
        "required": [
          "track",
          "primitives",
          "centreline"
        ],
        "title": "CorridorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Layout": {
        "enum": [
          "rwd",
          "awd",
          "fwd"
        ],
        "title": "Layout",
        "type": "string"
      },
...
      "PartSummary": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "connectors": {
            "title": "Connectors",
            "type": "integer"
          },
          "device": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "ldraw_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ldraw Id"
          },
          "mass_g": {
            "title": "Mass G",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "verified": {
            "title": "Verified",
            "type": "boolean"
          }
        },
        "required": [
          "key",
          "ldraw_id",
          "name",
          "category",
          "mass_g",
          "connectors",
          "device",
          "verified"
        ],
        "title": "PartSummary",
        "type": "object"
      },
      "Primitive": {
        "additionalProperties": false,
        "properties": {
          "color": {
            "title": "Color",
            "type": "string"
          },
          "kind": {
            "enum": [
              "box",
              "cylinder",
              "plane"
            ],
            "title": "Kind",
            "type": "string"
          },
          "opacity": {
            "default": 1.0,
            "title": "Opacity",
            "type": "number"
          },
          "pos": {
            "maxItems": 3,
            "minItems": 3,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "Pos",
            "type": "array"
          },
          "quat": {
            "maxItems": 4,
            "minItems": 4,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "Quat",
            "type": "array"
          },
          "size": {
            "maxItems": 3,
            "minItems": 3,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "Size",
            "type": "array"
          },
          "surface": {
            "default": "",
            "title": "Surface",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "pos",
          "quat",
          "size",
          "color"
        ],
        "title": "Primitive",
        "type": "object"
      },
      "QuickStartParams": {
        "additionalProperties": false,
        "description": "Inputs of the quick-start generator (spec 0002); invalid values list valid ones.",
        "properties": {
          "ackermann_pct": {
            "default": 60.0,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Ackermann Pct",
            "type": "number"
          },
          "battery": {
            "default": "powerbank",
            "enum": [
              "powerbank",
              "none"
            ],
            "title": "Battery",
            "type": "string"
          },
          "board": {
            "default": "raspberry_pi_5",
            "enum": [
              "raspberry_pi_5",
              "orange_pi_5",
              "none"
            ],
            "title": "Board",
            "type": "string"
          },
          "differential": {
            "default": true,
            "title": "Differential",
            "type": "boolean"
          },
          "drive_gears": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drive Gears"
          },
          "drive_motor": {
            "default": "ev3_large",
            "enum": [
              "ev3_large",
              "ev3_medium",
              "dc_motor"
            ],
            "title": "Drive Motor",
            "type": "string"
          },
          "layout": {
            "$ref": "#/components/schemas/Layout",
            "default": "rwd"
          },
          "max_steer_deg": {
            "default": 30.0,
            "maximum": 40.0,
            "minimum": 15.0,
            "title": "Max Steer Deg",
            "type": "number"
          },
          "measured_cog_mm": {
            "anyOf": [
              {
                "maxItems": 3,
                "minItems": 3,
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Measured Cog Mm"
          },
          "measured_mass_kg": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Measured Mass Kg"
          },
          "sensors": {
            "items": {
              "$ref": "#/components/schemas/SensorSpec"
            },
            "title": "Sensors",
            "type": "array"
          },
          "steering_motor": {
            "default": "ev3_medium",
            "enum": [
              "ev3_medium",
              "servo"
            ],
            "title": "Steering Motor",
            "type": "string"
          },
          "steering_play_deg": {
            "default": 3.0,
            "maximum": 15.0,
            "minimum": 0.0,
            "title": "Steering Play Deg",
            "type": "number"
          },
          "track_studs": {
            "default": 11,
            "title": "Track Studs",
            "type": "integer"
          },
          "wheel": {
            "default": "56908+55976",
            "title": "Wheel",
            "type": "string"
          },
          "wheelbase_studs": {
            "default": 15,
            "title": "Wheelbase Studs",
            "type": "integer"
          }
        },
        "title": "QuickStartParams",
        "type": "object"
      },
      "QuickstartResponse": {
        "additionalProperties": false,
        "properties": {
          "assembly": {
            "additionalProperties": true,
            "title": "Assembly",
            "type": "object"
          },
          "car": {
            "$ref": "#/components/schemas/CarScene"
          },
          "derived": {
            "additionalProperties": true,
            "title": "Derived",
            "type": "object"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/Warning"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "assembly",
          "derived",
          "warnings",
          "car"
        ],
        "title": "QuickstartResponse",
        "type": "object"
      },
...
      "ReplayRequest": {
        "additionalProperties": false,
        "properties": {
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "title": "ReplayRequest",
        "type": "object"
      },
      "ReplaySummary": {
        "additionalProperties": false,
        "properties": {
          "channels": {
            "additionalProperties": {
              "items": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "array"
            },
            "title": "Channels",
            "type": "object"
          },
          "duration_s": {
            "title": "Duration S",
            "type": "number"
          },
          "frames": {
            "title": "Frames",
            "type": "integer"
          },
          "speed_cmd": {
            "items": {
              "type": "number"
            },
            "title": "Speed Cmd",
            "type": "array"
          },
          "speed_meas": {
            "items": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Speed Meas",
            "type": "array"
          },
          "states": {
            "items": {
              "type": "string"
            },
            "title": "States",
            "type": "array"
          },
          "steering_cmd": {
            "items": {
              "type": "number"
            },
            "title": "Steering Cmd",
            "type": "array"
          },
          "t": {
            "items": {
              "type": "number"
            },
            "title": "T",
            "type": "array"
          },
          "truth_xy": {
            "items": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "number"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array"
            },
            "title": "Truth Xy",
            "type": "array"
          }
        },
        "required": [
          "frames",
          "duration_s",
          "t",
          "steering_cmd",
          "speed_cmd",
          "speed_meas",
          "states",
          "truth_xy",
          "channels"
        ],
        "title": "ReplaySummary",
        "type": "object"
      },
...
      "SceneBody": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "parts": {
            "items": {
              "$ref": "#/components/schemas/ScenePart"
            },
            "title": "Parts",
            "type": "array"
          }
        },
        "required": [
          "name",
          "parts"
        ],
        "title": "SceneBody",
        "type": "object"
      },
...
      "SensorSpec": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "enum": [
              "ev3_ultrasonic",
              "ev3_gyro",
              "ev3_touch",
              "lidar_2d"
            ],
            "title": "Kind",
            "type": "string"
          },
          "offset_mm": {
            "default": [
              0.0,
              0.0,
              0.0
            ],
            "maxItems": 3,
            "minItems": 3,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "Offset Mm",
            "type": "array"
          },
          "preset": {
            "enum": [
              "front",
              "left",
              "right",
              "rear",
              "center",
              "top"
            ],
            "title": "Preset",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "preset"
        ],
        "title": "SensorSpec",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "Warning": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "Warning",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "RaceForge engine",
    "version": "0.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/controllers": {
      "get": {
        "operationId": "controllers_api_v1_controllers_get",
        "parameters": [
          {
            "in": "query",
            "name": "extra",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Extra"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ControllerInfo"
                  },
                  "title": "Response Controllers Api V1 Controllers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Controllers"
      }
    },
    "/api/v1/corridor": {
      "post": {
        "operationId": "corridor_api_v1_corridor_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorridorParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorridorResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Corridor"
      }
    },
...
    "/api/v1/parts": {
      "get": {
        "operationId": "parts_api_v1_parts_get",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "default": "",
              "title": "Query",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "default": "",
              "title": "Category",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PartSummary"
                  },
                  "title": "Response Parts Api V1 Parts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Parts"
      }
    },
    "/api/v1/quickstart": {
      "post": {
        "operationId": "quickstart_api_v1_quickstart_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuickStartParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuickstartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Quickstart"
      }
    },
    "/api/v1/quickstart/export/{kind}": {
      "post": {
        "operationId": "export_api_v1_quickstart_export__kind__post",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "title": "Kind",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuickStartParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export"
      }
    },
...
    "/api/v1/replays": {
      "post": {
        "operationId": "replay_api_v1_replays_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplaySummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replay"
      }
    },
...
  }
}